Update and cleanup Celery integration (#2891)
Cleanup the Celery integration to make it easier to work on it and also update to the new minimum supported Celery versions and the new Scopes API.
